Updated User Login method for BestBooks APi object
Updates to login user method include while loop for user to enter in there passwords and username, if else statements that check if user name is correct and matches database, same for password. Option for returning to main menu is added if anything goes wrong. Created the start of a user menu