updated UserDatabase and BestBooksMain
changed the function in creationOfUsers from a statement to a prepared statement for better security and created a showAllUsers method for displaying the usernames of existing users.
^The above changes were done to the UseraDataBase object.
cahnged varible name of checkifuseraccountesits to newAccountUser. added else statement to signUp method; inside statement added a call to basicUserMenu. Realize i need to add a better way for user to go back to start menu incase they decide to change mind and not ccreate account. added tailrec type annotation to the start menu method and added import for tailrec.
^The above changes are for updates to BestBooksMain